# Conflicts:
#	test/SI_to_Conc/TestResults_SI2Conc_MJT_UoEdinburgh_num.csv
#	test/results/SI_to_Conc/TestResults_SI2Conc_LCB_BNI.csv
#	test/results/SI_to_Conc/TestResults_SI2Conc_LEK_UoEdinburgh.csv
#	test/results/SI_to_Conc/TestResults_SI2Conc_MB_UoManchester.csv
#	test/results/SI_to_Conc/TestResults_SI2Conc_MJT_UoEdinburgh.csv
#	test/results/SI_to_Conc/TestResults_SI2Conc_OG_MO_AUMC_ICR_RMH.csv
#	test/results/SI_to_Conc/TestResults_SI2Conc_ST_USydAus.csv
#	test/results/T1_mapping/TestResults_T1mapping_MJT_EdinburghUK_t1_VFA_lin.csv
#	test/results/T1_mapping/TestResults_T1mapping_MJT_EdinburghUK_t1_VFA_nonlin.csv

+++ src/original/MJT_UoEdinburghUK/dce_fit.py
@@ -63,7 +63,7 @@ class SigToEnh(Fitter):
         s_pre = np.mean(s[self.base_idx])
         if s_pre <= 0:
             raise ArithmeticError('Baseline signal is zero or negative.')
-        enh = np.empty(s.shape, dtype=np.float32)
+        enh = np.empty(s.shape, dtype=np.float64)
         enh[:] = 100. * ((s - s_pre) / s_pre) if s_pre > 0 else np.nan
         return enh
